overwrite old test with new one
