+++ tests/synthetic/test_compare_aurora_vs_archived_emtf.py
@@ -13,16 +13,25 @@ FORTRAN config file
 import numpy as np
 from pathlib import Path
 
+import pandas as pd
+from aurora.config.metadata.processing import Processing
+from aurora.config.config_creator import ConfigCreator
 from aurora.general_helper_functions import TEST_PATH
+from aurora.pipelines.helpers import initialize_config
 from aurora.pipelines.process_mth5 import process_mth5_run
+from aurora.pipelines.process_mth5_dev import process_mth5_from_dataset_definition
 from aurora.sandbox.io_helpers.zfile_murphy import read_z_file
-from aurora.test_utils.synthetic.make_processing_configs import create_test_run_config
+from aurora.test_utils.synthetic.make_processing_configs_new import create_test_run_config
+from aurora.tf_kernel.dataset import DatasetDefinition
+from aurora.tf_kernel.helpers import extract_run_summaries_from_mth5s
 from aurora.transfer_function.emtf_z_file_helpers import (
     merge_tf_collection_to_match_z_file,
 )
+from mth5.utils.helpers import initialize_mth5
 
 from make_mth5_from_asc import create_test1_h5
 from make_mth5_from_asc import create_test12rr_h5
+from plot_helpers_synthetic import plot_rho_phi
 
 SYNTHETIC_PATH = TEST_PATH.joinpath("synthetic")
 CONFIG_PATH = SYNTHETIC_PATH.joinpath("config")
@@ -31,21 +40,28 @@ EMTF_OUTPUT_PATH = SYNTHETIC_PATH.joinpath("emtf_output")
 AURORA_RESULTS_PATH = SYNTHETIC_PATH.joinpath("aurora_results")
 AURORA_RESULTS_PATH.mkdir(exist_ok=True)
 
-EXPECTED_RMS_MISFIT = {}
-EXPECTED_RMS_MISFIT["test1"] = {}
-EXPECTED_RMS_MISFIT["test1"]["rho"] = {}
-EXPECTED_RMS_MISFIT["test1"]["phi"] = {}
-EXPECTED_RMS_MISFIT["test1"]["rho"]["xy"] = 4.406358  #4.380757  # 4.357440
-EXPECTED_RMS_MISFIT["test1"]["phi"]["xy"] = 0.862902  #0.871609  # 0.884601
-EXPECTED_RMS_MISFIT["test1"]["rho"]["yx"] = 3.625859  #3.551043  # 3.501146
-EXPECTED_RMS_MISFIT["test1"]["phi"]["yx"] = 0.840394  #0.812733  # 0.808658
-EXPECTED_RMS_MISFIT["test2r1"] = {}
-EXPECTED_RMS_MISFIT["test2r1"]["rho"] = {}
-EXPECTED_RMS_MISFIT["test2r1"]["phi"] = {}
-EXPECTED_RMS_MISFIT["test2r1"]["rho"]["xy"] = 3.940519  #3.949857  #3.949919
-EXPECTED_RMS_MISFIT["test2r1"]["phi"]["xy"] = 0.959861  #0.962837  #0.957675
-EXPECTED_RMS_MISFIT["test2r1"]["rho"]["yx"] = 4.136467  #4.121772  #4.117700
-EXPECTED_RMS_MISFIT["test2r1"]["phi"]["yx"] = 1.635570  #1.637581  #1.629026
+def get_expected_rms_misfit(test_case_id, emtf_version=None):
+    expected_rms_misfit = {}
+    expected_rms_misfit["rho"] = {}
+    expected_rms_misfit["phi"] = {}
+    if test_case_id == "test1":
+        if emtf_version == "fortran":
+            expected_rms_misfit["rho"]["xy"] = 4.406358  #4.380757  # 4.357440
+            expected_rms_misfit["phi"]["xy"] = 0.862902  #0.871609  # 0.884601
+            expected_rms_misfit["rho"]["yx"] = 3.625859  #3.551043  # 3.501146
+            expected_rms_misfit["phi"]["yx"] = 0.840394  #0.812733  # 0.808658
+        elif emtf_version == "matlab":
+            expected_rms_misfit["rho"]["xy"] = 2.691072
+            expected_rms_misfit["phi"]["xy"] = 0.780713
+            expected_rms_misfit["rho"]["yx"] = 3.676269
+            expected_rms_misfit["phi"]["yx"] = 1.392265
+
+    elif test_case_id == "test2r1":
+        expected_rms_misfit["rho"]["xy"] = 3.940519  #3.949857  #3.949919
+        expected_rms_misfit["phi"]["xy"] = 0.959861  #0.962837  #0.957675
+        expected_rms_misfit["rho"]["yx"] = 4.136467  #4.121772  #4.117700
+        expected_rms_misfit["phi"]["yx"] = 1.635570  #1.637581  #1.629026
+    return expected_rms_misfit
 
 def compute_rms(rho, phi, model_rho_a=100.0, model_phi=45.0, verbose=False):
     """
@@ -78,122 +94,6 @@ def compute_rms(rho, phi, model_rho_a=100.0, model_phi=45.0, verbose=False):
         print(f"phi_rms = {phi_rms}")
     return rho_rms, phi_rms
 
-def make_subtitle(rho_rms_aurora, rho_rms_emtf,
-                  phi_rms_aurora, phi_rms_emtf,
-                  matlab_or_fortran, ttl_str=""):
-    """
-    
-    Parameters
-    ----------
-    rho_rms_aurora: float
-        rho_rms for aurora data differenced against a model. comes from compute_rms
-    rho_rms_emtf:
-        rho_rms for emtf data differenced against a model. comes from compute_rms
-    phi_rms_aurora:
-        phi_rms for aurora data differenced against a model. comes from compute_rms
-    phi_rms_emtf:
-        phi_rms for emtf data differenced against a model. comes from compute_rms
-    matlab_or_fortran: str
-        "matlab" or "fortran".  A specifer for the version of emtf.
-    ttl_str: str
-        string onto which we add the subtitle
-
-    Returns
-    -------
-    ttl_str: str
-        Figure title with subtitle
-
-    """
-    ttl_str += (
-        f"\n rho rms_aurora {rho_rms_aurora:.1f} rms_{matlab_or_fortran}"
-        f" {rho_rms_emtf:.1f}"
-    )
-    ttl_str += (
-        f"\n phi rms_aurora {phi_rms_aurora:.1f} rms_{matlab_or_fortran}"
-        f" {phi_rms_emtf:.1f}"
-    )
-    return ttl_str
-
-
-def make_figure_basename(local_station_id,
-                         reference_station_id,
-                         xy_or_yx,
-                         matlab_or_fortran):
-    """
-    
-    Parameters
-    ----------
-    local_station_id: str
-        station label
-    reference_station_id: str
-        remote reference station label
-    xy_or_yx: str
-        mode: "xy" or "yx"
-    matlab_or_fortran: str
-        "matlab" or "fortran".  A specifer for the version of emtf.
-
-    Returns
-    -------
-    figure_basename: str
-        filename for figure
-
-    """
-    station_string = f"{local_station_id}"
-    if reference_station_id:
-        station_string = f"{station_string}_rr{reference_station_id}"
-    figure_basename = (
-        f"synthetic_{station_string}_{xy_or_yx}_{matlab_or_fortran}.png"
-    )
-    return figure_basename
-
-def plot_rho_phi(xy_or_yx,
-                 tf_collection,
-                 rho_rms_aurora,
-                 rho_rms_emtf,
-                 phi_rms_aurora,
-                 phi_rms_emtf,
-                 matlab_or_fortran,
-                 aux_data=None,
-                 use_subtitle=True,
-                 show_plot=False):
-    """
-    Could be made into a method of TF Collection
-    Parameters
-    ----------
-    xy_or_yx
-    tf_collection
-    rho_rms_aurora
-    rho_rms_emtf
-    phi_rms_aurora
-    phi_rms_emtf
-    matlab_or_fortran
-    aux_data
-    use_subtitle
-    show_plot
-
-    Returns
-    -------
-
-    """
-    ttl_str = ""
-    if use_subtitle:
-        ttl_str = make_subtitle(rho_rms_aurora, rho_rms_emtf,
-                                phi_rms_aurora, phi_rms_emtf,
-                                matlab_or_fortran)
-
-    figure_basename = make_figure_basename(tf_collection.local_station_id,
-                                           tf_collection.reference_station_id,
-                                           xy_or_yx,
-                                           matlab_or_fortran)
-    tf_collection.rho_phi_plot(
-        aux_data=aux_data,
-        xy_or_yx=xy_or_yx,
-        ttl_str=ttl_str,
-        show=show_plot,
-        figure_basename=figure_basename,
-        figure_path=AURORA_RESULTS_PATH,
-    )
-    return
 
 
 def assert_rms_misfit_ok(expected_rms_misfit, xy_or_yx, rho_rms_aurora,
@@ -215,31 +115,35 @@ def assert_rms_misfit_ok(expected_rms_misfit, xy_or_yx, rho_rms_aurora,
     """
     expected_rms_rho = expected_rms_misfit['rho'][xy_or_yx]
     expected_rms_phi = expected_rms_misfit['phi'][xy_or_yx]
-    print(f"expected_rms_rho_xy {expected_rms_rho}")
-    print(f"expected_rms_rho_xy {expected_rms_phi}")
+    print(f"expected_rms_rho_{xy_or_yx} {expected_rms_rho}")
+    print(f"expected_rms_phi_{xy_or_yx} {expected_rms_phi}")
     assert np.isclose(rho_rms_aurora - expected_rms_rho, 0, atol=rho_tol)
     assert np.isclose(phi_rms_aurora - expected_rms_phi, 0, atol=phi_tol)
     return
 
 
 def process_synthetic_1_standard(
-        processing_config_path,
+        processing_config,
         auxilliary_z_file,
         z_file_base,
-        expected_rms_misfit=None,
+        dd_df = None,
+        expected_rms_misfit={},
         make_rho_phi_plot=True,
         show_rho_phi_plot=False,
         use_subtitle=True,
         emtf_version="matlab",
+        **kwargs
     ):
     """
 
     Parameters
     ----------
-    processing_config_path: str or Path
+    processing_config: str or Path, or a Processing() object
         where the processing configuration file is found
     expected_rms_misfit: dict
-        see description in assert_rms_misfit_ok
+        has expected values for the RMS misfits for the TF quantities rho_xy, rho_yx,
+        phi_xy, phi_yx. These are used to validate that the processing results don't
+        change much.
     make_rho_phi_plot
     show_rho_phi_plot
     use_subtitle
@@ -266,14 +170,41 @@ def process_synthetic_1_standard(
     -------
 
     """
+
     z_file_path = AURORA_RESULTS_PATH.joinpath(z_file_base)
 
-    run_id = "001"
-    tf_collection = process_mth5_run(
-        processing_config_path, run_id, units="MT", show_plot=False,
-        z_file_path=z_file_path
-    )
-    tf_collection._merge_decimation_levels()
+    cond1 = isinstance(processing_config, str)
+    cond2 = isinstance(processing_config, Path)
+    if (cond1 or cond2):
+        print("This needs to be updated to work with new mt_metadata Processing object")
+        #load from a json path or string
+        config = initialize_config(processing_config)
+    elif isinstance(processing_config, Processing):
+        config = processing_config
+        mth5_path = config.stations.local.mth5_path
+    else:
+        print(f"processing_config has unexpected type {type(processing_config)}")
+        raise Exception
+
+    #<get_channel_summary>
+    # mth5_obj = initialize_mth5(mth5_path, mode="r")
+    # ch_summary = mth5_obj.channel_summary
+    # dataset_definition = DatasetDefinition()
+    # dataset_definition.from_mth5_channel_summary(ch_summary)
+    # mth5_obj.close_mth5()
+
+    dataset_definition = DatasetDefinition()
+    dataset_definition.df = dd_df
+    #</get_channel_summary>
+
+
+
+
+    tf_collection = process_mth5_from_dataset_definition(config,
+                                               dataset_definition,
+                                               units="MT",
+                                               z_file_path=z_file_path)
+
 
     #END THE NORMAL PROCESSING TEST
 
@@ -305,33 +236,57 @@ def process_synthetic_1_standard(
                          emtf_version,
                          aux_data=aux_data,
                          use_subtitle=use_subtitle,
-                         show_plot=show_rho_phi_plot)
+                         show_plot=show_rho_phi_plot,
+                         output_path=AURORA_RESULTS_PATH)
     return
 
 
-def aurora_vs_emtf(test_case_id, emtf_version, auxilliary_z_file, z_file_base,
-                   expected_rms_misfit=None):
-    processing_config_path = create_test_run_config(test_case_id,
-                                                    matlab_or_fortran=emtf_version)
+def aurora_vs_emtf(test_case_id, emtf_version, auxilliary_z_file, z_file_base):
+    """
+
+    Parameters
+    ----------
+    test_case_id: str
+        one of ["test1", "test2r1"].  "test1" is associated with single station
+        processing. "test2r1" is remote refernce processing
+    emtf_version: str
+        one of ["fortran", "matlab"]
+    auxilliary_z_file: str or pathlib.Path
+        points to a .zss, .zrr or .zmm that EMTF produced that will be compared
+        against the python aurora output
+    z_file_base: str
+        This is the z_file that aurora will write its output to
+
+
+
+    Returns
+    -------
+
+    """
+    processing_config, dd_df = create_test_run_config(test_case_id,
+                                               matlab_or_fortran=emtf_version)
+    expected_rms_misfit = get_expected_rms_misfit(test_case_id, emtf_version)
+
     process_synthetic_1_standard(
-        processing_config_path,
+        processing_config,
         auxilliary_z_file,
         z_file_base,
+        dd_df=dd_df,
         expected_rms_misfit=expected_rms_misfit,
         emtf_version=emtf_version,
         make_rho_phi_plot=True,
         show_rho_phi_plot=False,
         use_subtitle=True,
     )
+    return
 
 
-def run_test1(emtf_version, expected_rms_misfit=None):
+def run_test1(emtf_version):
     print(f"Test1 vs {emtf_version}")
     test_case_id = "test1"
     auxilliary_z_file = EMTF_OUTPUT_PATH.joinpath("test1.zss")
     z_file_base = f"{test_case_id}_aurora_{emtf_version}.zss"
-    aurora_vs_emtf(test_case_id, emtf_version, auxilliary_z_file, z_file_base,
-                   expected_rms_misfit=expected_rms_misfit)
+    aurora_vs_emtf(test_case_id, emtf_version, auxilliary_z_file, z_file_base)
     return
 
 def run_test2r1():
@@ -340,15 +295,19 @@ def run_test2r1():
     emtf_version = "fortran"
     auxilliary_z_file = EMTF_OUTPUT_PATH.joinpath("test2r1.zrr")
     z_file_base = f"{test_case_id}_aurora_{emtf_version}.zrr"
-    aurora_vs_emtf(test_case_id, emtf_version, auxilliary_z_file, z_file_base,
-                   expected_rms_misfit=EXPECTED_RMS_MISFIT[test_case_id])
+    aurora_vs_emtf(test_case_id, emtf_version, auxilliary_z_file, z_file_base)
     return
+
+def make_mth5s():
+    mth5_path_1 = create_test1_h5()
+    mth5_path_2 = create_test12rr_h5()
+    return [mth5_path_1, mth5_path_2]
 
 
 def test():
-    create_test1_h5()
-    create_test12rr_h5()
-    run_test1("fortran", expected_rms_misfit=EXPECTED_RMS_MISFIT["test1"])
+    mth5_paths = make_mth5s()
+    super_summary = extract_run_summaries_from_mth5s(mth5_paths)
+    run_test1("fortran")
     run_test1("matlab")
     run_test2r1()
     print("success")
